+++ packages/zwave-js/src/lib/commandclass/index.ts
@@ -50,6 +50,8 @@ export type { MultilevelSwitchLevelChangeMetadata } from "./MultilevelSwitchCC";
 export type { NotificationMetadata } from "./NotificationCC";
 export { LocalProtectionState, RFProtectionState } from "./ProtectionCC";
 export { ToneId } from "./SoundSwitchCC";
+export { SupervisionStatus } from "./SupervisionCC";
+export type { SupervisionResult } from "./SupervisionCC";
 export { ThermostatMode } from "./ThermostatModeCC";
 export { ThermostatOperatingState } from "./ThermostatOperatingStateCC";
 export { SetbackType } from "./ThermostatSetbackCC";