+++ packages/zwave-js/src/lib/driver/Driver.ts
@@ -122,33 +122,33 @@ export interface ZWaveOptions {
 	/** Specify timeouts in milliseconds */
 	timeouts: {
 		/** how long to wait for an ACK */
-		ack: number;
+		ack: number; // >=1, default: 1000 ms
 		/** not sure */
-		byte: number;
+		byte: number; // >=1, default: 150 ms
 		/**
 		 * How long to wait for a controller response. Usually this timeout should never elapse,
 		 * so this is merely a safeguard against the driver stalling
 		 */
-		response: number;
+		response: number; // [500...5000], default: 1600 ms
 		/** How long to wait for a callback from the host for a SendData[Multicast]Request */
-		sendDataCallback: number;
+		sendDataCallback: number; // >=10000, default: 65000 ms
 		/** How much time a node gets to process a request and send a response */
-		report: number;
+		report: number; // [1000...40000], default: 1600 ms
 		/** How long generated nonces are valid */
-		nonce: number;
+		nonce: number; // [3000...20000], default: 5000 ms
 		/** How long a node is assumed to be awake after the last communication with it */
-		nodeAwake: number;
+		nodeAwake: number; // [1000...30000], default: 10000 ms
 	};
 
 	attempts: {
 		/** How often the driver should try communication with the controller before giving up */
-		controller: number;
+		controller: number; // [1...3], default: 3
 		/** How often the driver should try sending SendData commands before giving up */
-		sendData: number;
+		sendData: number; // [1...5], default: 3
 		/**
 		 * How many attempts should be made for each node interview before giving up
 		 */
-		nodeInterview: number;
+		nodeInterview: number; // [1...10], default: 5
 	};
 
 	/**
@@ -168,7 +168,7 @@ export interface ZWaveOptions {
 	/** Allows you to specify a different cache directory */
 	cacheDir: string;
 
-	/** Specify the network key to use for encryption */
+	/** Specify the network key to use for encryption. This must be a Buffer of exactly 16 bytes. */
 	networkKey?: Buffer;
 }
 
@@ -324,9 +324,12 @@ export interface SendMessageOptions {
 	 * Default: true
 	 */
 	changeNodeStatusOnMissingACK?: boolean;
-	/** Sets the number of milliseconds after which a transaction expires. When the expiration timer elapses, the transaction promise is rejected. */
+	/** Sets the number of milliseconds after which a message expires. When the expiration timer elapses, the promise is rejected with the error code `Controller_MessageExpired`. */
 	expire?: number;
-	/** Internal information used to identify or mark this transaction */
+	/**
+	 * Internal information used to identify or mark this transaction
+	 * @internal
+	 */
 	tag?: any;
 }
 
@@ -1218,7 +1221,7 @@ export class Driver extends EventEmitter {
 		log.driver.print("destroying driver instance...");
 
 		// First stop the send thread machine and close the serial port, so nothing happens anymore
-		this.sendThread.stop();
+		if (this.sendThread.initialized) this.sendThread.stop();
 		if (this.serial != undefined) {
 			if (this.serial.isOpen) await this.serial.close();
 			this.serial = undefined;